% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_version_yml.R
\name{create_log_session_info}
\alias{create_log_session_info}
\title{Log R session info}
\usage{
create_log_session_info(path_log = "R_sessionInfo.log")
}
\arguments{
\item{path_log}{Path to the file where the R session info log will
be written to. Default is the in the current directory as
\code{R_sessionInfo.log}.}
}
\value{
R session info log
}
\description{
This function logs the R session info to a file named \code{R_sessionInfo.log}
in the specified output directory.
}
\examples{
td <- withr::local_tempdir()
create_log_session_info(file.path(td, "session.log"))
}
