% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_version_yml.R
\name{create_versions_yml}
\alias{create_versions_yml}
\title{Create versions.yml file}
\usage{
create_versions_yml(packages, task_name, path_versions = "versions.yml")
}
\arguments{
\item{packages}{Named list of packages to add to  the
versions.yml. The items names should be the conda package
name, while the items value should be the package name
used in R.}

\item{task_name}{Name of the nextflow process.
Typically \verb{$\{task.process\}}}

\item{path_versions}{Path to the yml file where the versions will
be written to. Default is the in the current directory as
\code{versions.yml}}
}
\value{
versions.yml file
}
\description{
This function allows to automatically create the
versions.yml file used to store the packages used.
\code{r-base} as well as \code{r-nfcore.utils} versions will
automatically be added.
}
\examples{
td <- withr::local_tempdir()
create_versions_yml(
  list("r-stats" = "stats"), "MY_PROCESS",
  file.path(td, "my_versions.yml")
)
}
