% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is_null_or_invalid}
\alias{is_null_or_invalid}
\title{Check if is \code{NULL} or invalid}
\usage{
is_null_or_invalid(x, variable_name = NULL)
}
\arguments{
\item{x}{String to check}

\item{variable_name}{Name of the variable to print if error occurs}
}
\value{
TRUE if is null, ERROR if invalid and FALSE otherwise
}
\description{
Check if is \code{NULL} or invalid
}
\examples{
is_null_or_invalid(NULL) # TRUE
try(is_null_or_invalid(" ")) # ERROR
is_null_or_invalid("Hello World !") # FALSE
}
