% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_arguments.R
\name{is_valid_string}
\alias{is_valid_string}
\title{Check for Non-Empty, Non-Whitespace String}
\usage{
is_valid_string(input)
}
\arguments{
\item{input}{A variable to check.}
}
\value{
A logical value: TRUE if the input is a valid, non-empty,
non-whitespace string; FALSE otherwise.
}
\description{
This function checks if the input is non-NULL and contains more than
just whitespace.
It returns TRUE if the input is a non-empty, non-whitespace string,
and FALSE otherwise.
}
\examples{
is_valid_string("Hello World") # Returns TRUE
is_valid_string("   ")         # Returns FALSE
is_valid_string(NULL)          # Returns FALSE
}
