% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfcore.utils-package.R
\docType{package}
\name{nfcore.utils-package}
\alias{nfcore.utils}
\alias{nfcore.utils-package}
\title{The nf-core utils R package for Nextflow processes}
\description{
This package aims to ease the link between Nextflow inputs and
outputs to the R variable needed in the script.
}
\section{Functions}{

Below are listed some of the main functions available:

\code{\link[=process_inputs]{process_inputs()}}: Given a list of expected options with their
default this function will parse the arguments given and validate
their value based on the expected rules.

\code{\link[=process_end]{process_end()}}: Create a \code{versions.yml} and \code{R_sessionInfo.log}
file in the directory provided. The version file will be populated
with the R version, the version of nfcore.utils and the version of
the additional packages given.
}

\examples{
library(nfcore.utils)
td <- withr::local_tempdir()
test_file_path <- file.path(td, "test_file.txt")
file.create(test_file_path)
my_options <- list(
  input_file = test_file_path,
  output_file = "prefix",
  threshold = 0.5
)
args <- c("--threshold 0.7")
processed_options <- process_inputs(
  my_options, args,
  keys_to_nullify = c("input_file", "output_file"),
  expected_files = c("input_file"),
  expected_double = c("threshold"),
  required_opts = c("input_file", "threshold")
)
\dontrun{
process_end(
  packages = list(
    "r-stats" = "stats"
  ),
  task_name = 'MY_PROCESS'
)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nf-core/r-nf-core-utils}
  \item Report bugs at \url{https://github.com/nf-core/r-nf-core-utils/issues}
}

}
\author{
\strong{Maintainer}: Louis Le Nezet \email{louislenezet@gmail.com} (\href{https://orcid.org/0009-0000-0202-2703}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item nf-core Community
}

}
