% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nullify}
\alias{nullify}
\title{Turn “null” or empty strings into actual NULL}
\usage{
nullify(x)
}
\arguments{
\item{x}{Input option}
}
\value{
NULL or x
}
\description{
Turn “null” or empty strings into actual NULL
}
\examples{
nullify("null") # Returns NULL
nullify("NULL") # Returns NULL
nullify("   ")  # Returns NULL
nullify("")     # Returns NULL
nullify(NULL)   # Returns NULL
nullify(NA)     # Returns NULL
nullify("Hello") # Returns "Hello"
}
