% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_arguments.R
\name{parse_args}
\alias{parse_args}
\title{Parse out options from a string without recourse to optparse}
\usage{
parse_args(x)
}
\arguments{
\item{x}{Long-form argument list like --opt1 val1 --opt2 val2}
}
\value{
named list of options and values similar to optparse
}
\description{
Parse out options from a string without recourse to optparse
}
