% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_arguments.R
\name{parse_arguments}
\alias{parse_arguments}
\title{Parse out options from a string without recourse to optparse}
\usage{
parse_arguments(x)
}
\arguments{
\item{x}{Long-form argument list like --opt1 val1 --opt2 val2}
}
\value{
named list of options and values similar to optparse
}
\description{
If only the key is given, the value will be set to TRUE
}
\examples{
parse_arguments("--opt1 val1 --opt2 val2")
parse_arguments(' --opt1-extra "value with space" --opt2 val2 ')
}
