% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_inputs.R
\name{process_inputs}
\alias{process_inputs}
\title{Process input parameters and files}
\usage{
process_inputs(
  opt,
  args = NULL,
  keys_to_nullify = NULL,
  expected_files = NULL,
  expected_folders = NULL,
  expected_double = NULL,
  expected_integer = NULL,
  expected_boolean = NULL,
  required_opts = NULL
)
}
\arguments{
\item{opt}{A list of default options.}

\item{args}{A character vector of command-line arguments to override
defaults.}

\item{keys_to_nullify}{A character vector of keys that should be nullified.}

\item{expected_files}{A character vector of keys in \code{opt} that are
expected to be file paths.}

\item{expected_folders}{A character vector of keys in \code{opt} that are
expected to be folder paths.}

\item{expected_double}{A character vector of keys in \code{opt} that are
expected to be float values.}

\item{expected_integer}{A character vector of keys in \code{opt} that are
expected to be integer values.}

\item{expected_boolean}{A character vector of keys in \code{opt} that are
expected to be boolean values.}

\item{required_opts}{A character vector of keys in \code{opt} that are
required and must not be null or empty.}
}
\value{
A list of processed options with overrides applied and validated.
}
\description{
This function processes input parameters and files, applying parameter
overrides, validating expected keys and files, and ensuring that the
provided file paths are valid.
}
\examples{
td <- withr::local_tempdir()
test_file_path <- file.path(td, "test_file.txt")
file.create(test_file_path)
my_options <- list(
  input_file = test_file_path,
  output_file = "prefix",
  threshold = 0.5,
  "is-a-test" = NULL
)
args <- c("--threshold 0.7 --is-a-test")
processed_options <- process_inputs(
  my_options, args,
  keys_to_nullify = c("input_file", "output_file"),
  expected_files = c("input_file"),
  expected_double = c("threshold"),
  expected_boolean = c("is-a-test"),
  required_opts = c("input_file", "threshold")
)
}
