% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\name{read_delim_flexible}
\alias{read_delim_flexible}
\title{Flexibly read CSV or TSV files}
\usage{
read_delim_flexible(file, ...)
}
\arguments{
\item{file}{Input file}

\item{...}{
  Arguments passed on to \code{\link[utils:read.table]{utils::read.delim}}
  \describe{
    \item{\code{header}}{a logical value indicating whether the file contains the
    names of the variables as its first line.  If missing, the value is
    determined from the file format: \code{header} is set to \code{TRUE}
    if and only if the first row contains one fewer field than the
    number of columns.}
    \item{\code{quote}}{the set of quoting characters. To disable quoting
    altogether, use \code{quote = ""}.  See \code{\link{scan}} for the
    behaviour on quotes embedded in quotes.  Quoting is only considered
    for columns read as character, which is all of them unless
    \code{colClasses} is specified.}
    \item{\code{dec}}{the character used in the file for decimal points.}
    \item{\code{fill}}{logical. If \code{TRUE} then in case the rows have unequal
    length, blank fields are implicitly added.  See \sQuote{Details}.}
    \item{\code{comment.char}}{character: a character vector of length one
    containing a single character or an empty string.  Use \code{""} to
    turn off the interpretation of comments altogether.}
  }}
}
\value{
output Data frame
}
\description{
The following rules are applied:
\itemize{
\item extension == txt or tsv, then separator is set to tabulation
\item extension == csv, then separator is set to comma
}
}
