% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_string}
\alias{valid_string}
\title{Check for Non-Empty, Non-Whitespace String}
\usage{
valid_string(input)
}
\arguments{
\item{input}{A variable to check.}
}
\value{
\code{input} with whitespace trimmed or throw error if not a valid string
}
\description{
This function checks if the input is non-NULL and contains more than
just whitespace.
}
\examples{
valid_string("Hello World") # Returns "Hello World"
valid_string("  Hello World ") # Returns "Hello World"
valid_string("   ")         # Error
valid_string(NULL)          # Error
}
