% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_boolean}
\alias{validate_boolean}
\title{Check if value is a boolean}
\usage{
validate_boolean(x, variable_name = NULL)
}
\arguments{
\item{x}{String to check and convert}

\item{variable_name}{Name of the variable to print if error occurs}
}
\value{
x as TRUE/FALSE, or NULL if is null and ERROR if x
isn't convertable to a boolean
}
\description{
Convert to TRUE or FALSE with accepted lowered values as:
\itemize{
\item TRUE: 1, yes, true
\item FALSE: 0, no, false
}
}
\examples{
validate_boolean(NULL) # NULL
try(validate_boolean(" ")) # ERROR
try(validate_boolean("Hello World !")) # ERROR
try(validate_boolean(12.4)) # ERROR
validate_boolean("1") # TRUE
validate_boolean(0) # FALSE
}
