% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_double}
\alias{validate_double}
\title{Check if value is double number}
\usage{
validate_double(x, variable_name = NULL)
}
\arguments{
\item{x}{String to check and convert}

\item{variable_name}{Name of the variable to print if error occurs}
}
\value{
x as double, or NULL if is null and ERROR if x
isn't convertable to a double
}
\description{
Check if value is double number
}
\examples{
validate_double(NULL) # NULL
try(validate_double(" ")) # ERROR
try(validate_double("Hello World !")) # ERROR
validate_double("12.4") # 12.4
validate_double("12") # 12.0
}
