% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_file}
\alias{validate_file}
\title{Check if value is an existing file}
\usage{
validate_file(x, variable_name = NULL)
}
\arguments{
\item{x}{String to check}

\item{variable_name}{Name of the variable to print if error occurs}
}
\value{
x, or NULL if is null and ERROR if x
isn't an existing file
}
\description{
Check if value is an existing file
}
\examples{
try(validate_file("test")) # ERROR
}
