% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_folder}
\alias{validate_folder}
\title{Check if value is an existing folder}
\usage{
validate_folder(x, variable_name = NULL)
}
\arguments{
\item{x}{String to check}

\item{variable_name}{Name of the variable to print if error occurs}
}
\value{
x, or NULL if is null and ERROR if x
isn't an existing folder
}
\description{
Check if value is an existing folder
}
\examples{
try(validate_folder("test")) # ERROR
}
