% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_integer}
\alias{validate_integer}
\title{Check if value is an integer}
\usage{
validate_integer(x, variable_name = NULL)
}
\arguments{
\item{x}{String to check and convert}

\item{variable_name}{Name of the variable to print if error occurs}
}
\value{
x as integer, or NULL if is null and ERROR if x
isn't convertable to an integer
}
\description{
Check if value is an integer
}
\examples{
validate_integer(NULL) # NULL
try(validate_integer(" ")) # ERROR
try(validate_integer("Hello World !")) # ERROR
try(validate_integer("12.4")) # ERROR
validate_integer("12") # 12
}
