% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_version_yml.R
\name{create_log_session_info}
\alias{create_log_session_info}
\title{Log R session info}
\usage{
create_log_session_info(out_dir = ".")
}
\arguments{
\item{out_dir}{Output directory where the R session info log will be
written to. Default is the current directory.}
}
\value{
R session info log
}
\description{
This function logs the R session info to a file named \code{R_sessionInfo.log}
in the specified output directory.
}
\examples{
create_log_session_info("output_dir")
}
