% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_arguments.R
\name{nullify}
\alias{nullify}
\title{Turn “null” or empty strings into actual NULL}
\usage{
nullify(x)
}
\arguments{
\item{x}{Input option}
}
\value{
NULL or x
}
\description{
Turn “null” or empty strings into actual NULL
}
