% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_version_yml.R
\name{process_end}
\alias{process_end}
\title{Process end of the workflow}
\usage{
process_end(packages, task_name, out_dir = ".")
}
\arguments{
\item{packages}{Named list of packages to add to  the
versions.yml. The items names should be the conda package
name, while the items value should be the package name
used in R.}

\item{task_name}{Name of the nextflow process.
Typically \verb{$\{task.process\}}}

\item{out_dir}{Output directory where the versions.yml will
be written to. Default is the current directory.}
}
\value{
versions.yml file and R session info log
}
\description{
Create the versions.yml file and log session info at the end of the process.
}
\examples{
td <- withr::local_tempdir()
process_end(list("r-stats" = "stats"), "MY_PROCESS", td)
}
